CC_Project_init/train_seat_availability/requirements.txt
flask
waitress
requests
mysql-connector-python
flask-cors